candidate/requirements.txt
asgiref==3.8.1
class-registry==2.1.2
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
pillow==10.4.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.2
regex==2024.9.11
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
